manager: Hold onto an extra JobManagerMessage sender
Currently the only interfaces that use the job manager are Deck-specific, which
will cause the channel to be dropped early on other hardware. Since it exits
early, the manager sees this as an error condition and exits with a failure
status. By keeping an unused reference to it we can prevent it from being
dropped early.