inputplumber: Handle interface adding in spawned task
The receiver in the poller could fill up if several interfaces were added
before we finished testing one, which could cause the bus to deadlock. This
appears as one of the proxies hanging on a reply, but that's specifically
because the bus was backlogged waiting for the receiver to be drained. By
splitting the tasks, we can continue draining the receiver while waiting for
the proxy separately.